Process Play-by-Play Data to Track Lineups for Each Play

This function processes play-by-play data to track lineups. It adds details about the starting lineups, substitutions, and final lineups for each game period. The function also identifies free throw sequences and tracks the players involved in those events.

Usage

track_pbp_lineups(data, quarter_start_lineups, all_subs, fgs_data)

Arguments

data

A data frame with raw play-by-play data

quarter_start_lineups

A data frame with lineups to begin each quarter

all_subs

A data frame with all subs made

fgs_data

A data frame with fgs before and-one

Value

A data frame with the lineups for all play-by-play data

Details

The resulting data frame includes four additional columns:

lineup_away_pt

Tracks the total playtime for away lineups.

lineup_home_pt

Tracks the total playtime for home lineups.

lineup_away

Tracks the away lineup for calculating stats.

lineup_home

Tracks the home lineup for calculating stats.